+++ amazon/cognito/fake/src/main/kotlin/org/http4k/connect/amazon/cognito/FakeCognito.kt
@@ -1,7 +1,8 @@
 package org.http4k.connect.amazon.cognito
 
 import org.http4k.aws.AwsCredentials
-import org.http4k.connect.ChaosFake
+import org.http4k.chaos.ChaoticHttpHandler
+import org.http4k.chaos.start
 import org.http4k.connect.amazon.core.model.Region
 import org.http4k.core.Request
 import org.http4k.core.Response
@@ -11,7 +12,7 @@ import java.time.Clock.systemUTC
 
 class FakeCognito(
     private val clock: Clock = systemUTC()
-) : ChaosFake() {
+) : ChaoticHttpHandler() {
 
     override val app = { req: Request -> Response(OK).body("{}") }