+++ amazon/s3/fake/src/test/kotlin/org/http4k/connect/amazon/s3/RunningFakeS3GlobalTest.kt
@@ -1,8 +1,9 @@
 package org.http4k.connect.amazon.s3
 
+import org.http4k.chaos.defaultLocalUri
+import org.http4k.chaos.start
 import org.http4k.client.JavaHttpClient
 import org.http4k.connect.amazon.fakeAwsEnvironment
-import org.http4k.connect.defaultLocalUri
 import org.http4k.core.then
 import org.http4k.filter.ClientFilters.SetHostFrom
 import org.http4k.server.Http4kServer