+++ amazon/sts/fake/src/main/kotlin/org/http4k/connect/amazon/sts/FakeSTS.kt
@@ -1,7 +1,8 @@
 package org.http4k.connect.amazon.sts
 
 import org.http4k.aws.AwsCredentials
-import org.http4k.connect.ChaosFake
+import org.http4k.chaos.ChaoticHttpHandler
+import org.http4k.chaos.start
 import org.http4k.connect.amazon.core.model.Region
 import org.http4k.core.Method.POST
 import org.http4k.routing.bind
@@ -13,7 +14,7 @@ import java.time.Duration.ofHours
 class FakeSTS(
     private val clock: Clock = Clock.systemUTC(),
     defaultSessionValidity: Duration = ofHours(1)
-) : ChaosFake() {
+) : ChaoticHttpHandler() {
 
     override val app = routes(
         "/" bind POST to routes(