+++ core/fake/src/main/kotlin/org/http4k/connect/deprecatedCoreFake.kt
@@ -1,0 +1,6 @@
+package org.http4k.connect
+
+import org.http4k.chaos.ChaoticHttpHandler
+
+@Deprecated("replaced with http4k type", ReplaceWith("org.http4k.chaos.ChaoticHttpHandler"))
+typealias ChaosFake = ChaoticHttpHandler