+++ core/fake/src/test/kotlin/org/http4k/connect/FakeSystemContract.kt
@@ -5,6 +5,9 @@ import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.containsSubstring
 import com.natpryce.hamkrest.equalTo
 import com.natpryce.hamkrest.greaterThan
+import org.http4k.chaos.ChaoticHttpHandler
+import org.http4k.chaos.defaultLocalUri
+import org.http4k.chaos.defaultPort
 import org.http4k.core.ContentType.Companion.APPLICATION_JSON
 import org.http4k.core.Method.GET
 import org.http4k.core.Request
@@ -16,7 +19,7 @@ import org.http4k.hamkrest.hasContentType
 import org.http4k.hamkrest.hasStatus
 import org.junit.jupiter.api.Test
 
-abstract class FakeSystemContract(private val fake: ChaosFake) {
+abstract class FakeSystemContract(private val fake: ChaoticHttpHandler) {
     protected abstract val anyValid: Request
 
     @Test